docs: support anchors in <details> (#832)
This builds on ideas in @marcelm's #831 by keeping the front page
streamlined while still supporting anchors that can be used as
permalinks.

First, the `details_ext.py` Sphinx extension was updated to support
adding an optional `:anchor:` argument. Using this option causes an
anchor to be created at the top of the `<details>` element, which
repeats the summary text followed by a `¶` with the permalink.

Then there's additional JS in the `docs/source/templates/layout.html`
that, upon visiting an anchor, looks for the closest `<details>` to the
target and makes it visible.

So once built, the link
https://bioconda.github.io/index.html#do-not-modify-condarc will link
to, and open if folded, the details section about how `-c bioconda` is
not sufficient (the original inspiration for #831).