docs: remove docs (#839)
This PR removes the documentation from this repo to be instead hosted
over on [bioconda-docs](https://github.com/bioconda/bioconda-docs).

The idea is to decouple the back-end build infrastructure for
bioconda-recipes (the primary goal of this bioconda-utils repo) from the
public-facing static docs. Hopefully it will allow documentation to be
iterated on faster (much faster tests, no release management required,
hopefully easier to contribute to).