feat: add ability to store build failures as yaml next to recipe, add flag to automatically do so upon build failures, consider such files as blacklisting if they include `blacklist: true`. (#888)
This will enable a more automated bulk update process, where we
auto-blacklist failing recipes that are not libraries.
We can then organize fixing those blacklisted recipes in a follow-up
hackathon. For that purpose, I plan a bot command that automatically
generates and updates an issue that lists all blacklisted recipes sorted
by their download count in descending order (for prioritization). People
can then choose one and post a link to the PR in which they try to fix
it.