fix: adds alternate lookup if base lookup fails (#896)
This is a first pass attempt at fixing #895.

Main changes are updating the GithubRelease hoster to look for alternate
paths via the GitHub API. The API sends a JSON response (example:
https://api.github.com/repos/PacificBiosciences/HiPhase/releases) that
contains each release along with a list of assets. The assets are
searched for files that match the provides pattern. Currently, this
first pass puts this as a backup to the main one (which I think is
always failing now). I also did not update the tests, I think someone
may have a better idea of how to do that. Happy to help push this
further, but wanted to get the ball rolling.

---------

Co-authored-by: Johannes Köster <johannes.koester@tu-dortmund.de>