Recover --docker-base-image feature
In the https://github.com/bioconda/bioconda-utils/pull/866 we introduced the `--docker-base-image` to help user to specify customized build docker image (such as Linux aarch64 build).

But we noticed that it doesn’t work after the https://github.com/bioconda/bioconda-utils/pull/894 , because a local var introduced and the `docker-base-image` arguments is ignored:
https://github.com/bioconda/bioconda-utils/blame/9a85115ae306f58c8b4e65e5f92f6cbdb5b68f04/bioconda_utils/cli.py#L468

This PR try to recover it, if user specify the `--docker-base-image`, it will be used directly otherwise use the local var.

Signed-off-by: Yikun Jiang <yikunkero@gmail.com>