feat: Bulk build failure wiki (#948)
In order to have a `bulk` branch-specific Build Failures wiki page, need
to allow for a git range to be specified for which recipes will be
included. This uses the same method of determining whether a recipe is
included as the build command. (i.e., it's not diffing the build failure
files between branches.)

Also added handling for recipes that exist but have no corresponding
package in the channel. That avoids errors from new Bioconductor recipes
during a bulk update.