feat: Bioconductor skeleton updates (#1027)
- When running `bioconda-utils bioconductor-skeleton` on Bioconductor
data packages, I was getting self-signed SSL errors (@bgruening is this
expected?). This change allow such errors to be ignored, returning None
for the url.
- In this round of Bioconductor 3.20 updates, it seems like a lot of
packages are now missing `zlib`. I still haven't pinpointed the reason,
but I'm guessing it has something to do with the updates to pinnings. So
for BioC packages that are detected to be compiled, this adds `zlib` to
host dependencies.