fix: keep existing patches during bioconductor updates (#1041)
Currently when doing a bulk Bioconductor upgrade, the patches section
has been lost. That leads to errors and manually digging through commits
to see what was lost. If we keep the patches, even if they need
updating, at least it will be obvious.

Also set `sort_keys=False` so that top-level yaml items are written in
the traditional order instead of being alphabetized. (Diffs will be
annoying for the next upgrade but, I think future readability is worth
it.)

I confirmed that the `additional-platforms` section in `extra` _is_
getting kept through upgrades, so no changes there.